fix(JAISIU-2156): Phase 0 close — restore DecideEngine back-compat + HumanBytes int64 cast after reorder
The cherry-pick of fix/JAISIU-2157-backcompat-empty-reason (which
combined two pre-Phase-0 contract fixes — DecideEngine empty string for
GGML choices, and uint64→int64 for format.HumanBytes) got auto-merged
into the original changes, which dropped them.

Re-applies both in the close branch. Verified locally via:
    bash scripts/test-operator-env.sh
    go test ./runner/ -count=1
    go test ./server/ -run TestPrismalama -count=1
    go test -tags=integration -short ./integration -run "TestShip|TestPrismalama" -count=1
    make sync-audit-check
    make print-defaults
All green.

Refs: JAISIU-2156, JAISIU-2157, JAISIU-2158.

+++ server/prismalama_capabilities.go
@@ -122,10 +122,11 @@ func buildPrismalamaCapabilities() api.PrismalamaCapabilitiesResponse {
 	resp.Environment.PRISMALAMA_AIRLLM_PYTHONPATH = os.Getenv("PRISMALAMA_AIRLLM_PYTHONPATH")
 
 	// Resolved numeric env vars (parsed). Falls back to envconfig defaults if env unset.
+	// Cast uint64 → int64 for format.HumanBytes (which expects int64).
 	resp.Resolved.GpuOverheadBytes = envconfig.GpuOverhead()
-	resp.Resolved.GpuOverheadHuman = format.HumanBytes(envconfig.GpuOverhead())
+	resp.Resolved.GpuOverheadHuman = format.HumanBytes(int64(envconfig.GpuOverhead()))
 	resp.Resolved.StreamingBudgetBytes = envconfig.StreamingBudgetBytes()
-	resp.Resolved.StreamingBudgetHuman = format.HumanBytes(envconfig.StreamingBudgetBytes())
+	resp.Resolved.StreamingBudgetHuman = format.HumanBytes(int64(envconfig.StreamingBudgetBytes()))
 
 	// Build info — GoVersion is filled at runtime; CompiledAt is best-effort
 	// (operators can wire it via -ldflags -X .../version.CompiledAt=<RFC3339>).
